refactor(python): alphabetize package list
This will make it easier to scan for existing packages and determine where to place addition packages in the future.

Aside: In order to build the protobuf docs, I considered adding the [protobuf-compiler package](https://packages.ubuntu.com/xenial/protobuf-compiler), but it's quite out-of-date in Ubuntu 16.04. I'll follow-up with a second image with C++ build tools available, instead.