fix: Disable build of the Ruby images
The canonical Ruby images are actually in googleapis/google-cloud-ruby (and the images currently in this repo are out of date and can break Ruby CI if they are built). Disable upload of these images for now to prevent that. We still need to figure out how to reconcile the two sets of images.