fix: Temporarily disable build of the Ruby images (#74)
The canonical Ruby images are actually elsewhere, in https://github.com/googleapis/google-cloud-ruby. The images currently in testing-infra-docker are out of date and can break Ruby CI if they are built. Disable upload of these images for now to prevent that. We still need to figure out how to reconcile the two sets of images.